Fix launching with Siri (#54)
The app sometimes would launch and just not process any audio input.

I found I was able to consistently reproduce this when launching the app
via Siri: "Hey Siri, launch Zen Tuner".

To fix this, I periodically check for the microphone access and then
start the audio capture. I use an exponential backoff strategy so this
happens as quickly as possible without over-taxing the system.

As part of this change, Zen Tuner no longer stops the audio capture
explicitly when backgrounding the app. The OS takes care of doing this
automatically in my testing (iOS 16.2).